Revert "[FIX] Перенести запрет на удаление отдела в слой сервиса"
This reverts commit f97d4182bec1e971466f04267986bc31f415b3c4.

+++ src/main/java/project/controller/SubdepartmentController.java
@@ -17,6 +17,7 @@ import static project.exceptions.Validator.validateDepartmentId;
 public class SubdepartmentController {
 
     private final SubdepartmentServiceImpl service;
+    private final StudentServiceImpl studentService;
 
     @GetMapping("/dep_{N}/subdep/data")
     public List<Subdepartment> getAll(@PathVariable("N") String depId,
@@ -47,6 +48,9 @@ public class SubdepartmentController {
                                         @PathVariable("id") short id,
                                         @RequestHeader(value = HttpHeaders.AUTHORIZATION, defaultValue = "") String jwtToken) {
         Integer departmentId = validateDepartmentId(depId);
+        if (studentService.getStudentExistsWithSubdepartmentId(departmentId, id)) {
+            throw new IllegalArgumentException("Существует работник из этого отдела. Необходимо добавить его в другой отдел.");
+        }
         service.deleteById(departmentId, id);
     }
